+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - PixPullSubscription.parse method
 - CreditSigner.resendToken method
+- Pdf and Payment methods to CreditNote resource
+- Update method to IssuingPurchase resource
+### Fixed
+- MerchantCategoryNumber type in IssuingPurchase resource, the API returns a string
 
 ## [1.2.0] - 2026-07-03
 ### Fixed